move cli functions into dedicated module
